drop databases
lastFile:ui/src/pages/RagChatTab/FooterComponents/RagChatQueryInput.tsx
