+++ ui/src/pages/RagChatTab/FooterComponents/RagChatQueryInput.tsx
@@ -36,7 +36,7 @@
  * DATA.
  ******************************************************************************/
 
-import { Button, Flex, Input, InputRef, Modal, Tooltip } from "antd";
+import { Button, Flex, Input, InputRef, Tooltip } from "antd";
 import {
   DatabaseFilled,
   DatabaseOutlined,
@@ -61,9 +61,7 @@ import { cdlBlue600, cdlRed600 } from "src/cuix/variables.ts";
 import { useSuggestQuestions } from "src/api/ragQueryApi.ts";
 import SuggestedQuestionsFooter from "pages/RagChatTab/FooterComponents/SuggestedQuestionsFooter.tsx";
 import ToolsManagerButton from "pages/RagChatTab/FooterComponents/ToolsManager.tsx";
-import DocumentationIcon from "src/cuix/icons/DocumentationIcon.ts";
-import useModal from "src/utils/useModal.ts";
-import UploadedFilesTable from "pages/DataSources/ManageTab/UploadedFilesTable.tsx";
+import SessionDocuments from "pages/RagChatTab/FooterComponents/SessionDocuments.tsx";
 
 const { TextArea } = Input;
 
@@ -92,7 +90,6 @@ const RagChatQueryInput = ({
     strict: false,
   });
   const inputRef = useRef<InputRef>(null);
-  const documentModal = useModal();
   const {
     data: sampleQuestions,
     isFetching: sampleQuestionsIsFetching,
@@ -169,36 +166,6 @@ const RagChatQueryInput = ({
     streamChatMutation.reset();
   };
 
-  const SessionDocuments = () => (
-    <>
-      <Tooltip title={"Add documents to Session"}>
-        <Button
-          size="small"
-          type="text"
-          onClick={() => {
-            documentModal.setIsModalOpen(true);
-          }}
-          icon={<DocumentationIcon style={{ color: cdlBlue600 }} />}
-        />
-      </Tooltip>
-      <Modal
-        title="Session Documents"
-        open={documentModal.isModalOpen}
-        footer={null}
-        onCancel={() => {
-          documentModal.setIsModalOpen(false);
-        }}
-        destroyOnHidden={true}
-        width={800}
-      >
-        <UploadedFilesTable
-          dataSourceId={activeSession.associatedDataSourceId.toString()}
-          summarizationModel={activeSession.inferenceModel}
-          simplifiedTable={true}
-        />
-      </Modal>
-    </>
-  );
   return (
     <div>
       <Flex vertical align="center" gap={10}>
@@ -287,9 +254,8 @@ const RagChatQueryInput = ({
                       onClick={handleExcludeKnowledgeBase}
                     />
                   </Tooltip>
-                  {activeSession?.associatedDataSourceId ? (
-                    <SessionDocuments />
-                  ) : null}
+
+                  <SessionDocuments activeSession={activeSession} />
                   <Button
                     size="small"
                     type="text"
